requirements.txt
# 基本依赖
flask==2.0.1
flask-socketio==5.1.1
python-dotenv==0.19.0
requests==2.26.0
werkzeug==2.0.1

# AI模型接口
openai==0.27.0

# 测试工具
pytest==6.2.5

# 开发工具
black==21.8b0
flake8==3.9.2
